requirements.txt
# API and Web Framework
fastapi==0.109.0
uvicorn==0.25.0
pydantic==2.5.3
starlette==0.36.1
python-dotenv==1.0.0
python-multipart==0.0.6
json-logging==1.3.0

# Database Drivers
redis==5.0.1
pymongo==4.6.1
neo4j==5.14.0
qdrant-client==1.7.0
kafka-python==2.0.2

# Utilities
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.3.2
tqdm==4.66.1
loguru==0.7.2

# Vector Embeddings
scipy==1.11.4
scikit-learn==1.3.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Type Checking
mypy==1.8.0
types-redis==5.0.0.2
types-requests==2.31.0.20231231
